+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Markdown docs generation using typedoc
  - `absoluteUrl` function
  - `hash` function
+ - `parseBool` function
 
 ### Removed
  - `conf.json`